+++ lib/src/services/CommandableHttpService.dart
@@ -1,4 +1,3 @@
-import 'dart:convert';
 import 'package:angel_framework/angel_framework.dart' as angel;
 import 'package:pip_services3_commons/pip_services3_commons.dart';
 import './RestService.dart';