+++ lib/src/services/HttpEndpoint.dart
@@ -1,5 +1,4 @@
 import 'dart:async';
-import 'dart:convert';
 
 import 'package:angel_framework/angel_framework.dart' as angel;
 import 'package:angel_framework/http.dart';